miraikugua/pyproject.toml
[tool.pdm]
distribution = false

[tool.pdm.build]
includes = []
[project]
name = "miraikugua"
version = "0.1.0"
description = ""
authors = [
    {name = "momordica2020", email = "287859992@qq.com"},
]
dependencies = [
    "graia-ariadne<1.0.0,>=0.11.7",
    "graia-saya>=0.0.19",
    "saucenao-api>=2.3.1",
    "pathlib>=1.0.1",
    "arclet-alconna-graia>=0.14.0",
    "arclet-alconna-ariadne>=0.14.0",
    "graiax-shortcut>=0.3.0",
]
requires-python = "<4.0,>=3.12"
readme = "README.md"
license = {text = "AGPL-3.0"}

